further refactoring of the parachain module
