+++ parachain/src/lib.rs
@@ -37,15 +37,14 @@
 //!              ^~~returned pointer
 //! ```
 //!
-//! The `load_params` and `write_result` functions provide utilities for setting up
-//! a parachain WASM module in Rust.
+//! The `wasm_api` module (enabled only with the wasm-api feature) provides utilities
+//!  for setting up a parachain WASM module in Rust.
 
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(not(feature = "std"), feature(alloc))]
 
 /// Re-export of parity-codec.
 pub extern crate parity_codec as codec;
-extern crate polkadot_primitives;
 
 #[macro_use]
 extern crate parity_codec_derive;
@@ -65,17 +64,17 @@ extern crate error_chain;
 
 #[cfg(not(feature = "std"))]
 use alloc::vec::Vec;
-use codec::{Encode, Decode};
 
 #[cfg(feature = "std")]
 pub mod wasm_executor;
 
-pub use polkadot_primitives::parachain::Id as ParaId;
+#[cfg(feature = "wasm-api")]
+pub mod wasm_api;
 
 /// Validation parameters for evaluating the parachain validity function.
 // TODO: consolidated ingress and balance downloads
-#[derive(PartialEq, Eq, Encode)]
-#[cfg_attr(feature = "std", derive(Debug, Decode))]
+#[derive(PartialEq, Eq, Decode)]
+#[cfg_attr(feature = "std", derive(Debug, Encode))]
 pub struct ValidationParams {
 	/// The collation body.
 	pub block_data: Vec<u8>,
@@ -97,7 +96,7 @@ pub struct ValidationResult {
 #[cfg_attr(feature = "std", derive(Debug, Decode))]
 pub struct Message {
 	/// The target parachain for this message.
-	pub target: ParaId,
+	pub target: u32,
 	/// The message data.
 	pub data: Vec<u8>,
 }
@@ -106,36 +105,8 @@ pub struct Message {
 #[cfg(feature = "std")]
 pub struct MessageRef<'a> {
 	/// The target parachain.
-	pub target: ParaId,
+	pub target: u32,
 	/// Underlying data of the message.
 	pub data: &'a [u8],
 }
 
-/// Load the validation params from memory when implementing a Rust parachain.
-///
-/// Offset and length must have been provided by the validation
-/// function's entry point.
-pub unsafe fn load_params(offset: usize, len: usize) -> ValidationParams {
-	let mut slice = ::core::slice::from_raw_parts(offset as *const u8, len);
-
-	ValidationParams::decode(&mut slice).expect("Invalid input data")
-}
-
-/// Allocate the validation result in memory, getting the return-pointer back.
-///
-/// As described in the crate docs, this is a pointer to the appended length
-/// of the vector.
-pub fn write_result(result: ValidationResult) -> usize {
-	let mut encoded = result.encode();
-	let len = encoded.len();
-
-	assert!(len <= u32::max_value() as usize, "Len too large for parachain-WASM abi");
-	(len as u32).using_encoded(|s| encoded.extend(s));
-
-	// do not alter `encoded` beyond this point. may reallocate.
-	let end_ptr = &encoded[len] as *const u8 as usize;
-
-	// leak so it doesn't get zeroed.
-	::core::mem::forget(encoded);
-	end_ptr
-}
